move projections stuff out of fold.rs into its own file
